apps/table/requirements.txt
# Table inference service - docling-ibm-models TFPredictor (TableFormer)
docling-ibm-models[opencv-python-headless]>=3.11.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
prometheus-client>=0.21.0
pillow>=10.0.0
huggingface-hub>=0.23.0
